JSON decoding methods for the Request type
This is NOT good: the request method should be payload agnostic.
Instead, define the decoding methods in the handlers.

Moreover, decoding JSON is handler specific, so it should not be
middleware in the first place.

Still commiting this because I want to save it somewhere. But the plan
is never to merge this into master. Rather, delete all this cruft and
the stuff Vince put in and go back to basics.